third_party/WebKit/LayoutTests/fast/dom/custom/constructor-calls-created-synchronously.html

<!DOCTYPE html>
<script src="../../js/resources/js-test-pre.js"></script>
<div id="container"></div>
<script>
description('Tests that custom element constructors run the createdCallback synchronously.');

var proto = Object.create(HTMLElement.prototype);
var ncallbacks = 0;
proto.createdCallback = function () {
  ncallbacks++;
};
var A = document.register('x-a', {prototype: proto});
var x = new A();
shouldBe('ncallbacks', '1');

successfullyParsed = true;
</script>
